Print method for rcpptimer output

Prints the times object and scales the timings if appropriate. If all timings are smaller than 1 microsecond, the timings are printed in nanoseconds. If the smallest timing is higher than a Millisecond / Seconds / Minutes / Hours, the timings are printed in the unit of that threshold. This behavior can be disabled by setting scale = FALSE.

Usage

# S3 method for class 'rcpptimer'
print(x, scale = TRUE, ...)

Arguments

x

Object of class rcpptimer

scale

Scale the timings and statistics to a more human readable format

...

further arguments are ignored